+++ superfreetts_addon/component_settings.py
@@ -13,6 +13,7 @@ from . import component_changes
 from . import component_donation
 from . import component_preferences
 from . import component_services
+from . import component_usage
 from . import config_models
 from . import gui_utils
 from . import i18n
@@ -25,7 +26,7 @@ class SettingsDialog(aqt.qt.QDialog):
     def __init__(self, hypertts, initial_tab=0, parent=None):
         super().__init__(parent)
         self.hypertts = hypertts
-        self.initial_tab = 0 if initial_tab not in (0, 1, 2, 3, 4) else initial_tab
+        self.initial_tab = 0 if initial_tab not in (0, 1, 2, 3, 4, 5) else initial_tab
         self._saved_once = False
         self._initial_snapshot = None
 
@@ -41,6 +42,7 @@ class SettingsDialog(aqt.qt.QDialog):
         self.changes_page = component_changes.ChangesPage(hypertts)
         self.donation_page = component_donation.DonationPage(hypertts)
         self.about_page = component_about.AboutPage(hypertts)
+        self.usage_page = component_usage.UsagePage(hypertts)
 
         self._page_buttons = []
         self._page_built = set()
@@ -82,12 +84,14 @@ class SettingsDialog(aqt.qt.QDialog):
         self.btn_changes = self._create_sidebar_button(i18n.get_text("tab_changes", lang))
         self.btn_donation = self._create_sidebar_button(i18n.get_text("tab_donation", lang))
         self.btn_about = self._create_sidebar_button(i18n.get_text("tab_about", lang))
+        self.btn_usage = self._create_sidebar_button(i18n.get_text("tab_usage", lang))
         self._page_buttons = [
             self.btn_services,
             self.btn_preferences,
             self.btn_changes,
             self.btn_donation,
             self.btn_about,
+            self.btn_usage,
         ]
 
         for btn in self._page_buttons:
@@ -110,11 +114,13 @@ class SettingsDialog(aqt.qt.QDialog):
         self.changes_container = self._build_page_container()
         self.donation_container = self._build_page_container()
         self.about_container = self._build_page_container()
+        self.usage_container = self._build_page_container()
         self.pages.addWidget(self.services_container)
         self.pages.addWidget(self.preferences_container)
         self.pages.addWidget(self.changes_container)
         self.pages.addWidget(self.donation_container)
         self.pages.addWidget(self.about_container)
+        self.pages.addWidget(self.usage_container)
         body_layout.addWidget(self.pages, 1)
 
         main_layout.addLayout(body_layout, 1)
@@ -139,6 +145,7 @@ class SettingsDialog(aqt.qt.QDialog):
         self.btn_changes.clicked.connect(lambda: self.switch_page(2))
         self.btn_donation.clicked.connect(lambda: self.switch_page(3))
         self.btn_about.clicked.connect(lambda: self.switch_page(4))
+        self.btn_usage.clicked.connect(lambda: self.switch_page(5))
 
         self.switch_page(self.initial_tab)
 
@@ -175,6 +182,8 @@ class SettingsDialog(aqt.qt.QDialog):
             self.donation_page.draw(self.donation_container.layout())
         elif index == 4:
             self.about_page.draw(self.about_container.layout())
+        elif index == 5:
+            self.usage_page.draw(self.usage_container.layout())
         self._page_built.add(index)
 
     def switch_page(self, index):